+++ packages/zowe-explorer/src/Profiles.ts
@@ -54,7 +54,7 @@ let InputBoxOptions: vscode.InputBoxOptions;
 export class Profiles extends ProfilesCache {
     // Processing stops if there are no profiles detected
     public static async createInstance(log: zowe.imperative.Logger): Promise<Profiles> {
-        Profiles.loader = new Profiles(log, vscode.workspace.workspaceFolders?.[0].uri.fsPath);
+        Profiles.loader = new Profiles(log, vscode.workspace.workspaceFolders?.[0]?.uri.fsPath);
         await Profiles.loader.refresh(ZoweExplorerApiRegister.getInstance());
         return Profiles.loader;
     }
@@ -640,7 +640,7 @@ export class Profiles extends ProfilesCache {
             let user = false;
             let global = true;
             let rootPath = getZoweDir();
-            if (vscode.workspace.workspaceFolders) {
+            if (vscode.workspace.workspaceFolders && vscode.workspace.workspaceFolders[0]) {
                 const choice = await this.getConfigLocationPrompt("create");
                 if (choice === undefined) {
                     vscode.window.showInformationMessage(
@@ -649,7 +649,7 @@ export class Profiles extends ProfilesCache {
                     return;
                 }
                 if (choice === "project") {
-                    rootPath = vscode.workspace.workspaceFolders?.[0].uri.fsPath;
+                    rootPath = vscode.workspace.workspaceFolders[0].uri.fsPath;
                     user = true;
                     global = false;
                 }
@@ -672,7 +672,7 @@ export class Profiles extends ProfilesCache {
                 homeDir: getZoweDir(),
                 projectDir: getFullPath(rootPath),
             });
-            if (vscode.workspace.workspaceFolders) {
+            if (vscode.workspace.workspaceFolders && vscode.workspace.workspaceFolders[0]) {
                 config.api.layers.activate(user, global, rootPath);
             }
 
@@ -714,24 +714,6 @@ export class Profiles extends ProfilesCache {
             vscode.window.showErrorMessage(
                 localize("Profiles.getProfileInfo.error", "Error in creating team configuration file: {0}", err.message)
             );
-        }
-    }
-
-    public updateImperativeSettings() {
-        try {
-            const fileName = path.join(getZoweDir(), "settings", "imperative.json");
-            const updatedSettings = {
-                overrides: {
-                    CredentialManager: false,
-                },
-            };
-            fs.writeFile(fileName, JSON.stringify(updatedSettings), "utf8", (err) => {
-                if (err) {
-                    this.log.error("Could not update imperative.json settings file", err);
-                }
-            });
-        } catch (error) {
-            this.log.error(error);
         }
     }
 
@@ -1463,7 +1445,7 @@ export class Profiles extends ProfilesCache {
         const existingLayers: zowe.imperative.IConfigLayer[] = [];
         const config = await zowe.imperative.Config.load("zowe", {
             homeDir: getZoweDir(),
-            projectDir: vscode.workspace.workspaceFolders?.[0].uri.fsPath,
+            projectDir: vscode.workspace.workspaceFolders?.[0]?.uri.fsPath,
         });
         const layers = config.layers;
         layers.forEach((layer) => {
@@ -1917,7 +1899,6 @@ export class Profiles extends ProfilesCache {
             for (const profile of Object.entries(newConfig.profiles)) {
                 delete newConfig.profiles[profile[0]].secure;
             }
-            this.updateImperativeSettings();
             newConfig.autoStore = false;
         }
     }