+++ packages/zowe-explorer/src/Profiles.ts
@@ -54,7 +54,7 @@ let InputBoxOptions: vscode.InputBoxOptions;
 export class Profiles extends ProfilesCache {
     // Processing stops if there are no profiles detected
     public static async createInstance(log: zowe.imperative.Logger): Promise<Profiles> {
-        Profiles.loader = new Profiles(log, vscode.workspace.workspaceFolders?.[0].uri.fsPath);
+        Profiles.loader = new Profiles(log, vscode.workspace.workspaceFolders?.[0]?.uri.fsPath);
         await Profiles.loader.refresh(ZoweExplorerApiRegister.getInstance());
         return Profiles.loader;
     }
@@ -640,7 +640,7 @@ export class Profiles extends ProfilesCache {
             let user = false;
             let global = true;
             let rootPath = getZoweDir();
-            if (vscode.workspace.workspaceFolders) {
+            if (vscode.workspace.workspaceFolders && vscode.workspace.workspaceFolders[0]) {
                 const choice = await this.getConfigLocationPrompt("create");
                 if (choice === undefined) {
                     vscode.window.showInformationMessage(
@@ -649,7 +649,7 @@ export class Profiles extends ProfilesCache {
                     return;
                 }
                 if (choice === "project") {
-                    rootPath = vscode.workspace.workspaceFolders?.[0].uri.fsPath;
+                    rootPath = vscode.workspace.workspaceFolders[0].uri.fsPath;
                     user = true;
                     global = false;
                 }
@@ -672,7 +672,7 @@ export class Profiles extends ProfilesCache {
                 homeDir: getZoweDir(),
                 projectDir: getFullPath(rootPath),
             });
-            if (vscode.workspace.workspaceFolders) {
+            if (vscode.workspace.workspaceFolders && vscode.workspace.workspaceFolders[0]) {
                 config.api.layers.activate(user, global, rootPath);
             }
 
@@ -1451,7 +1451,7 @@ export class Profiles extends ProfilesCache {
         const existingLayers: zowe.imperative.IConfigLayer[] = [];
         const config = await zowe.imperative.Config.load("zowe", {
             homeDir: getZoweDir(),
-            projectDir: vscode.workspace.workspaceFolders?.[0].uri.fsPath,
+            projectDir: vscode.workspace.workspaceFolders?.[0]?.uri.fsPath,
         });
         const layers = config.layers;
         layers.forEach((layer) => {