+++ packages/zowe-explorer-api/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to the "zowe-explorer-api" extension will be documented in this file.
 
+## TBD Release
+
+### Bug fixes
+
+- Fixed `ZoweVsCodeExtension` failing to initialize in environment with empty workspace. [#1994](https://github.com/zowe/vscode-extension-for-zowe/issues/1994)
+
 ## `2.4.0`
 
 ## TBD Release